Windows package version: adapt to identifiers that are SHAs
prerelease_identifier is now a short commit SHA,
so for the Windows installer version, which requires
an A.B.C.D version format with all positive integers,
we will set it the identifier to 99.

Note that previously Concourse was setting it to 0
unconditionally and no one has noticed that
in many years.

So that version can be considered to be irrelevant
for preview releases, and in GA releases,
the final part will be a .0 regardless
of the release.